Fix project srtucture and rename files
