+++ src/app/components/chat/chat.module.ts
@@ -1,19 +1,19 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import {ChatComponent} from './chat.component';
+import {MessagesComponent} from './messages/messages.component';
 import { routes } from './chat.routing';
-import { MaterialModule } from '../material.module';
+import { MaterialModule } from '../../material.module';
 import { ChatEffects } from './store/chat.effects';
 import { EffectsModule } from '@ngrx/effects';
 import { chatReducer } from './store/chat.reducer';
 import { StoreModule } from '@ngrx/store';
-import {ChatService} from '../services/chat.service';
+import {ChatService} from '../../services/chat.service';
 import {HttpClientModule} from '@angular/common/http';
 
 @NgModule({
   declarations: [
-    ChatComponent
+    MessagesComponent
   ],
   imports: [
     BrowserModule,
@@ -26,7 +26,7 @@ import {HttpClientModule} from '@angular/common/http';
   providers: [
     ChatService
   ],
-  bootstrap: [ChatComponent]
+  bootstrap: [MessagesComponent]
 })
 export class ChatModule {
 }
