+++ src/app/components/chat/store/chat.effects.spec.ts
@@ -2,10 +2,10 @@ import { cold, hot } from 'jasmine-marbles';
 import { ChatEffects } from './chat.effects';
 import { TestBed } from '@angular/core/testing';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { ChatService } from '../../services/chat.service';
+import { ChatService } from '../../../services/chat.service';
 import { TestHotObservable } from 'jasmine-marbles/src/test-observables';
 import { provideMockActions } from '@ngrx/effects/testing';
-import { Message } from '../../models/message.interface';
+import { Message } from '../../../models/message.interface';
 import {of, throwError} from 'rxjs';
 import * as ChatActions from './chat.actions';
 import { StoreModule } from '@ngrx/store';
